requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
sentence-transformers==2.5.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
pydantic==2.6.4
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
